+++ psutil/_pswindows.py
@@ -80,6 +80,7 @@ __extra__all__ = [
 
 CONN_DELETE_TCB = "DELETE_TCB"
 HAS_PROC_IO_PRIORITY = hasattr(cext, "proc_io_priority_get")
+HAS_GET_LOADAVG = hasattr(cext, "getloadavg")
 
 
 if enum is None:
@@ -351,6 +352,21 @@ def cpu_freq():
     curr, max_ = cext.cpu_freq()
     min_ = 0.0
     return [_common.scpufreq(float(curr), min_, float(max_))]
+
+
+if HAS_GET_LOADAVG:
+    _loadavg_inititialized = False
+
+    def getloadavg():
+        global _loadavg_inititialized
+
+        if not _loadavg_inititialized:
+            cext.init_loadavg_counter()
+            _loadavg_inititialized = True
+
+        # Drop to 2 decimal points which is what Linux does
+        raw_loads = cext.getloadavg()
+        return tuple([round(load, 2) for load in raw_loads])
 
 
 # =====================================================================