+++ psutil/tests/test_system.py
@@ -38,6 +38,7 @@ from psutil.tests import enum
 from psutil.tests import get_test_subprocess
 from psutil.tests import HAS_BATTERY
 from psutil.tests import HAS_CPU_FREQ
+from psutil.tests import HAS_GET_LOADAVG
 from psutil.tests import HAS_NET_IO_COUNTERS
 from psutil.tests import HAS_SENSORS_BATTERY
 from psutil.tests import HAS_SENSORS_FANS
@@ -777,6 +778,11 @@ class TestSystemAPIs(unittest.TestCase):
 
         if LINUX:
             self.assertEqual(len(ls), psutil.cpu_count())
+
+    @unittest.skipIf(not HAS_GET_LOADAVG, "not supported")
+    def test_getloadavg(self):
+        loadavg = psutil.getloadavg()
+        assert len(loadavg) == 3
 
     def test_os_constants(self):
         names = ["POSIX", "WINDOWS", "LINUX", "MACOS", "FREEBSD", "OPENBSD",